+++ @utahdts/utah-design-system/react/components/forms/MultiSelect/MultiSelectTags.jsx
@@ -51,7 +51,8 @@ export function MultiSelectTags({ isDisabled }) {
                               addPoliteMessage,
                               selectedValueIndex,
                               selectedOption,
-                              multiSelectContextNonStateRef
+                              multiSelectContextNonStateRef,
+                              multiSelectContext.onChange
                             );
                           })
                         )