+++ @utahdts/utah-design-system/react/components/forms/MultiSelect/context/MultiSelectContextProvider.jsx
@@ -63,7 +63,6 @@ export default function MultiSelectContextProvider({
         }
       });
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
     [values]
   );