+++ @utahdts/utah-design-system/react/components/table/hooks/useTableFilterRegistration.js
@@ -9,10 +9,12 @@ import { useTableContext } from './useTableContext';
  * each TableFilterX component needs to setup and register its basic information. The context has context
  * global level filter settings, but each filter has its own settings, like `exactMatch`. This hook sets
  * up context filter information for a filter
+ * @template TableDataT
  * @param {string} recordFieldPath path to the field in the record on which this filter applies
  * @param {boolean} exactMatch should the filter be an exact match on the data
+ * @param {TableDataT} defaultValue filter can have a default starting value
  */
-export function useTableFilterRegistration(recordFieldPath, exactMatch) {
+export function useTableFilterRegistration(recordFieldPath, exactMatch, defaultValue) {
   const { setState } = useTableContext();
 
   useEffect(
@@ -21,7 +23,7 @@ export function useTableFilterRegistration(recordFieldPath, exactMatch) {
       setState((draftState) => {
         draftState.filterValues.value[recordFieldPath] = {
           exactMatch,
-          value: '',
+          value: defaultValue,
         };
       });